Ref: 샘플 데이터 초기화 로직 개선
